+++ src/cli/note/delete.rs
@@ -1,15 +1,21 @@
 use meilisearch_sdk::documents::DocumentDeletionQuery;
-use sea_orm::{ColumnTrait, EntityTrait, ModelTrait, QueryFilter, TransactionTrait};
+use sea_orm::{ActiveModelTrait, ColumnTrait, EntityTrait, FromQueryResult, ModelTrait, QueryFilter, QuerySelect, TransactionTrait};
+use serde::{Deserialize, Serialize};
 
 use crate::{
   config::{load_config, IdMethod},
   consts::MeiliNotes,
   db::postgres::connect_pg,
-  entities::note,
+  entities::{note, sea_orm_active_enums::NoteVisibilityEnum},
   util::id::{
     aid::gen_aid, aidx::gen_aidx, meid::gen_meid, objectid::gen_object_id, ulid::gen_ulid,
   },
 };
+
+#[derive(Debug, Serialize, Deserialize, FromQueryResult)]
+struct NoteIdModel {
+  id: String,
+}
 
 // ノートを削除するコマンド
 // 1. 対象となるノートを検索する
@@ -20,6 +26,11 @@ pub async fn delete(
   config_path: &str,
   host: Option<&str>,
   days: u64,
+  visibility: Option<Vec<NoteVisibilityEnum>>,
+  no_reaction: bool,
+  no_reply: bool,
+  no_renote: bool,
+  no_clipped: bool,
 ) -> Result<(), Box<dyn std::error::Error>> {
   // read config for judging whether meilisearch is enabled
   let config = load_config(config_path).unwrap();
@@ -30,65 +41,97 @@ pub async fn delete(
     .with_max_level(tracing::Level::INFO)
     .init();
 
-  // do if meilisearch is enabled
-  if let Some(meili_config) = config.meilisearch {
-    // get meilisearch config
-    let meili_host = meili_config.host;
-    let port = meili_config.port;
-    let ssl = meili_config.ssl;
-    let api_key = meili_config.api_key;
-    let m_index = meili_config.index;
-    tracing::info!("Meilisearch is enabled");
+  let should_delete_from_meilisearch = visibility
+    .as_ref()
+    .map(|v| {
+      !v.iter()
+        .any(|vis| matches!(vis, NoteVisibilityEnum::Public | NoteVisibilityEnum::Home))
+    })
+    .unwrap_or(true);
 
-    // format meilisearch index name
-    let uid = format!("{}---notes", m_index);
+  if should_delete_from_meilisearch {
+    // do if meilisearch is enabled
+    if let Some(meili_config) = config.meilisearch {
+      // get meilisearch config
+      let meili_host = meili_config.host;
+      let port = meili_config.port;
+      let ssl = meili_config.ssl;
+      let api_key = meili_config.api_key;
+      let m_index = meili_config.index;
+      tracing::info!("Meilisearch is enabled");
 
-    // parse meilisearch url for meilisearch_sdk
-    let url = format!(
-      "{}://{}:{}",
-      if ssl { "https" } else { "http" },
-      meili_host,
-      port
-    );
-    // create meilisearch client
-    let client = meilisearch_sdk::client::Client::new(url, Some(api_key)).unwrap();
-    // createdAtをfilterの条件としてmeilisearchから削除する
-    let index = client.index(uid);
-    let now = std::time::SystemTime::now()
-      .duration_since(std::time::UNIX_EPOCH)
-      .unwrap()
-      .as_secs();
-    let date = now - days * 24 * 60 * 60;
-    if date == 0 {
-      tracing::error!("Cannot specify today or future date");
-      std::process::exit(3);
+      // format meilisearch index name
+      let uid = format!("{}---notes", m_index);
+
+      // parse meilisearch url for meilisearch_sdk
+      let url = format!(
+        "{}://{}:{}",
+        if ssl { "https" } else { "http" },
+        meili_host,
+        port
+      );
+      // create meilisearch client
+      let client = meilisearch_sdk::client::Client::new(url, Some(api_key)).unwrap();
+
+      // create pg client for meilisearch
+      let pg_client_meili = connect_pg(config_path).await.unwrap();
+      let mut pg_query = note::Entity::find().select_only().column(note::Column::Id);
+      // generate pg query for options
+      if no_reaction {
+        pg_query = pg_query.filter(note::Column::Reactions.ne("{}"));
+      }
+      if no_reply {
+        pg_query = pg_query.filter(note::Column::RepliesCount.eq(0));
+      }
+      if no_renote {
+        pg_query = pg_query.filter(note::Column::RenoteCount.eq(0));
+      }
+      if no_clipped {
+        pg_query = pg_query.filter(note::Column::ClippedCount.eq(0));
+      }
+
+      let ids: Vec<String> = pg_query
+        .into_model::<NoteIdModel>()
+        .all(&pg_client_meili)
+        .await?
+        .into_iter()
+        .map(|note| note.id)
+        .collect();
+
+      // createdAtをfilterの条件としてmeilisearchから削除する
+      let index = client.index(uid);
+      let now = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .unwrap()
+        .as_secs();
+      let date = now - days * 24 * 60 * 60;
+
+      let created_at = format!("createdAt < {}", date);
+      let mut filters = vec![created_at.clone()];
+      if let Some(host) = host {
+        filters.push(format!("userHost = \"{}\"", host));
+      } else {
+        filters.push("userHost IS NOT NULL".to_string());
+      }
+
+      if !ids.is_empty() {
+        let id_filter = format!("id IN [{}]", ids.join(","));
+        filters.push(id_filter);
+      }
+      let filter = filters.join(" AND ");
+
+      // do query
+      let task = DocumentDeletionQuery::new(&index)
+        .with_filter(&filter)
+        .execute::<MeiliNotes>()
+        .await?;
+
+      task.wait_for_completion(&client, None, None).await.unwrap();
+      tracing::info!("Meilisearch delete task completed");
+      pg_client_meili.close().await?;
     }
-
-    let created_at = format!("createdAt < {}", date);
-    let base_filter = created_at.clone();
-
-    let task = if let Some(host) = host {
-      let user_host = format!("userHost = \"{}\"", host);
-      let query = format!("{} AND {}", base_filter, user_host);
-      DocumentDeletionQuery::new(&index)
-        .with_filter(&query)
-        .execute::<MeiliNotes>()
-        .await
-        .unwrap()
-    } else {
-      let user_host = "userHost IS NOT NULL".to_string();
-      let query = format!("{} AND {}", base_filter, user_host);
-      DocumentDeletionQuery::new(&index)
-        .with_filter(&query)
-        .execute::<MeiliNotes>()
-        .await
-        .unwrap()
-    };
-
-    task.wait_for_completion(&client, None, None).await.unwrap();
-    tracing::info!("Meilisearch delete task completed");
   } else {
-    tracing::info!("Meilisearch is not enabled");
+    tracing::info!("Meilisearch is not enabled or specified visibility is not including public or home, skipping");
   }
 
   // start transaction
@@ -101,16 +144,12 @@ pub async fn delete(
   } else {
     query = query.filter(note::Column::UserHost.is_not_null());
   }
-  // TODO: ここでIDから日付を取得してlteにsubstrを入れてクエリを生成する
+
   let now = std::time::SystemTime::now()
     .duration_since(std::time::UNIX_EPOCH)
     .unwrap()
     .as_millis() as u64;
   let date = now - days * 24 * 60 * 60;
-  if date == 0 {
-    tracing::error!("Cannot specify today or future date");
-    std::process::exit(3);
-  }
 
   match config.id {
     IdMethod::Aid => {
@@ -135,8 +174,53 @@ pub async fn delete(
     }
   }
 
+  if let Some(visibility) = visibility {
+    query = query.filter(note::Column::Visibility.is_in(visibility));
+  }
+
+  if no_reaction {
+    query = query.filter(note::Column::Reactions.ne("{}"));
+  }
+
+  if no_reply {
+    query = query.filter(note::Column::RepliesCount.eq(0));
+  }
+
+  if no_renote {
+    query = query.filter(note::Column::RenoteCount.eq(0));
+  }
+
+  if no_clipped {
+    query = query.filter(note::Column::ClippedCount.eq(0));
+  }
+
   // delete from database
   let notes = query.all(&txn).await?;
+  let mut reply_ids = std::collections::HashSet::new();
+  for note in &notes {
+    if let Some(reply_id) = &note.reply_id {
+      reply_ids.insert(reply_id.clone());
+    }
+  }
+
+  if !reply_ids.is_empty() {
+    let reply_ids_vec: Vec<String> = reply_ids.into_iter().collect();
+
+    let reply_targets = note::Entity::find()
+    .filter(note::Column::Id.is_in(reply_ids_vec))
+    .all(&txn)
+    .await?;
+
+    for target in reply_targets {
+      let mut target_model: note::ActiveModel = target.clone().into();
+
+      let new_count = if target.replies_count > 0 { target.replies_count - 1 } else { 0 };
+      target_model.replies_count = sea_orm::ActiveValue::Set(new_count);
+
+      target_model.update(&txn).await?;
+    }
+  }
+
   let chunk_size = 100;
   for chunk in notes.chunks(chunk_size) {
     let futures = chunk.iter().map(|note| note.clone().delete(&txn));