+++ src/cli/user/mod.rs
@@ -2,7 +2,8 @@ pub mod cull;
 pub mod delete;
 pub mod prune;
 
-use clap::{Parser, Subcommand};
+use crate::consts::UserIdentifier;
+use clap::{ArgGroup, Parser, Subcommand};
 use cull::cull;
 use delete::delete;
 use prune::prune;
@@ -20,11 +21,30 @@ pub enum UserSubCommand {
     #[arg(short = 'c', long = "config", default_value = ".config/default.yml")]
     config_path: String,
   },
+  #[command(group(
+    ArgGroup::new("user_delete")
+      .required(true)
+      .args(&["id", "username"])
+  ),
+    about = "Delete user account(s)",
+    long_about = "Delete remote user account(s) by ID or username. Need specify either ID or username.")]
   Delete {
     #[arg(short = 'c', long = "config", default_value = ".config/default.yml")]
     config_path: String,
-    #[arg(short = 'i', long = "id", value_delimiter = ',')]
-    id: Vec<String>,
+    #[arg(
+      short = 'i',
+      long = "id",
+      value_delimiter = ',',
+      help = "Id of user you want to delete. You can specify multiple IDs with comma. e.g. -i a4ct2ps000,a3tghk0000"
+    )]
+    id: Option<Vec<String>>,
+    #[arg(
+      short = 'u',
+      long = "username",
+      value_delimiter = ',',
+      help = "Username of user you want to delete. You can specify multiple usernames with comma. e.g. -u alice,bob"
+    )]
+    user: Option<Vec<UserIdentifier>>,
   },
   Prune {
     #[arg(short = 'c', long = "config", default_value = ".config/default.yml")]
@@ -44,9 +64,15 @@ impl UserCommand {
       UserSubCommand::Cull { config_path } => {
         cull(config_path).await?;
       }
-      UserSubCommand::Delete { config_path, id } => {
-        let refs = id.iter().map(|s| s.as_str()).collect();
-        delete(config_path, refs).await?;
+      UserSubCommand::Delete {
+        config_path,
+        id,
+        user,
+      } => {
+        let id_ref = id
+          .as_ref()
+          .map(|v| v.iter().map(|s| s.as_str()).collect::<Vec<&str>>());
+        delete(config_path, id_ref, user.clone()).await?;
       }
       UserSubCommand::Prune {
         config_path,