+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.2.0
+
+## Release Date
+2025/03/02
+
+## Feat
+- `user delete`で`username`を指定できるように
+- `note delete`で`visibility`を指定できるように
+- `note delete`で反応がないノートのみを削除できるように
+
+## Fix
+
 # 0.1.3
 
 ## Release Date